requirements.txt
plotly
jinja2
tqdm
pywebview